+++ project/printf/printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: shachowd <shachowd@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 09:38:21 by shachowd          #+#    #+#             */
-/*   Updated: 2024/07/03 20:43:38 by shachowd         ###   ########.fr       */
+/*   Updated: 2024/07/03 21:11:45 by shachowd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ int	ft_printpercent(void);
 int	ft_printstr(char *str);
 int	ft_printnum(int n);
 int	ft_print_unsigned(int nbr);
-int ft_print_ptr(unsigned long ptr);
+int ft_print_ptr(unsigned long long ptr);
 int ft_print_hex(unsigned int n, const char format);
 int	ft_printf(const char *str, ...);
